Refactor some code to reduce duplication
